Move IndexAction from client to service
